test: 로또 등수 관련 테스트코드 변경
