+++ src/test/java/lotto/domain/LottosTest.java
@@ -9,7 +9,8 @@ class LottosTest {
     @Test
     @DisplayName("구매할 갯수를 올바르게 구하는지")
     void Calculate_Lotto_Count() {
-        Lottos lottos = new Lottos(new Money("10000",1000));
+        int unitPrice = 1000;
+        Lottos lottos = new Lottos(new Money("10000", unitPrice), unitPrice);
         assertThat(lottos.getLottosSize()).isEqualTo(10);
     }
 }
